Dead code removal, renames and code motion for readability
